requirements.txt
Django>=5.2.7
djangorestframework>=3.16.1
psycopg2-binary>=2.9.9
beautifulsoup4>=4.12.0
selenium>=4.15.0
webdriver-manager>=4.0.1
dateparser>=1.2.0
requests>=2.31.0
gunicorn>=21.2.0
